// web/src/api/models/CreateIdentityRequest.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
import type { Value } from './Value';
export type CreateIdentityRequest = {
    attributes?: Value;
    display_name?: string | null;
    login: string;
    password?: string | null;
};