// web/src/api/models/PermissionAssignmentResponse.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type PermissionAssignmentResponse = {
    created: string;
    id: number;
    identity_id: number;
    permission_set_id: number;
    permission_set_ref: string;
};